Revert "[APPSEC-56614] Add Instrumentation::Gateway#pushed? method"
