+++ sig/datadog/appsec/instrumentation/gateway.rbs
@@ -7,17 +7,26 @@ module Datadog
         type final_call = ^() -> final_call_result
         type stack = ::Proc
 
+        # TODO: this should be a struct
         type stack_result = ::Array[nil | middleware_result | final_call_result]
+
+        class Middleware
+          attr_reader key: ::Symbol
+          attr_reader block: ::Proc
+
+          def initialize: (::Symbol key) { (stack next, Instrumentation::Gateway::Argument env) -> stack_result } -> void
+          def call: (stack next, Gateway::Argument? env) -> stack_result
+        end
 
         @middlewares: ::Hash[::String, ::Array[Middleware]]
 
         def initialize: () -> void
-
         def push: (::String name, Gateway::Argument env) ?{ () -> final_call_result } -> stack_result
-
         def watch: (::String name, ::Symbol key) { (stack next, Gateway::Argument env) -> stack_result } -> void
 
-        def pushed?: (::String name) -> bool
+        private
+
+        attr_reader middlewares: ::Hash[::String, ::Array[Middleware]]
       end
 
       self.@gateway: Gateway
