--- sig/datadog/appsec/instrumentation/gateway/middleware.rbs
@@ -1,15 +1,0 @@
-module Datadog
-  module AppSec
-    module Instrumentation
-      class Gateway
-        class Middleware
-          attr_reader key: ::Symbol
-          attr_reader block: ::Proc
-
-          def initialize: (::Symbol key) { (Gateway::stack next, Gateway::Argument env) -> stack_result } -> void
-          def call: (Gateway::stack next, Gateway::Argument? env) -> stack_result
-        end
-      end
-    end
-  end
-end
