+++ spec/datadog/appsec/instrumentation/gateway_spec.rb
@@ -5,7 +5,7 @@ require 'datadog/appsec/instrumentation/gateway'
 
 RSpec.describe Datadog::AppSec::Instrumentation::Gateway do
   subject(:gateway) { described_class.new }
-  let(:middlewares) { gateway.instance_variable_get(:@middlewares) }
+  let(:middlewares) { gateway.send(:middlewares) }
 
   describe '#watch' do
     it 'stores middleware' do
@@ -79,21 +79,6 @@ RSpec.describe Datadog::AppSec::Instrumentation::Gateway do
       expect(result[0][1]).to eq(:done)
       expect(env_1).to eq({ a: :b })
       expect(env_2).to eq({ a: :b, c: :d })
-    end
-  end
-
-  describe '#pushed?' do
-    it { expect(gateway.pushed?('event.0')).to be(false) }
-
-    it 'returns true if event was pushed' do
-      expect { gateway.push('event.1', {}) }.to change { gateway.pushed?('event.1') }
-        .from(false).to(true)
-
-      expect { gateway.push('event.2', {}) }.to change { gateway.pushed?('event.2') }
-        .from(false).to(true)
-
-      expect { gateway.push('event.2', {}) }.not_to change { gateway.pushed?('event.2') }
-        .from(true)
     end
   end
 end
